@@ -18,63 +18,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.0.4",
-    "@angular/common": "^15.0.4",
-    "@angular/compiler": "^15.0.4",
-    "@angular/core": "^15.0.4",
-    "@angular/forms": "^15.0.4",
-    "@angular/localize": "^15.0.4",
-    "@angular/platform-browser": "^15.0.4",
-    "@angular/platform-browser-dynamic": "^15.0.4",
-    "@angular/router": "^15.0.4",
-    "@ngxs/router-plugin": "^3.7.6",
-    "@ngxs/storage-plugin": "^3.7.6",
-    "@ngxs/store": "^3.7.6",
+    "@angular/animations": "^16.0.0",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+    "@angular/localize": "^16.0.0",
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/router": "^16.0.0",
+    "@ngxs/router-plugin": "^3.8.1",
+    "@ngxs/storage-plugin": "^3.8.1",
+    "@ngxs/store": "^3.8.1",
     "fast-json-patch": "^3.0.0-1",
     "moment": "^2.29.0",
-    "primeflex": "~2.0.0",
+    "primeflex": "^3.3.0",
     "primeicons": "^6.0.1",
-    "primeng": "^15.0.1",
-    "rxjs": "~6.6.3",
+    "primeng": "^16.0.0",
+    "rxjs": "^7.8.1",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.5"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^15.0.0",
-    "@angular-devkit/build-angular": "^15.0.5",
-    "@angular-eslint/builder": "^15.1.0",
-    "@angular-eslint/eslint-plugin": "^15.1.0",
-    "@angular-eslint/eslint-plugin-template": "^15.1.0",
-    "@angular-eslint/schematics": "^15.1.0",
-    "@angular-eslint/template-parser": "^15.1.0",
-    "@angular/cli": "^15.0.5",
-    "@angular/compiler-cli": "^15.0.4",
-    "@angular/language-service": "^15.0.4",
-    "@badeball/cypress-cucumber-preprocessor": "^15.1.5",
+    "@angular-builders/custom-webpack": "^16.0.0",
+    "@angular-devkit/build-angular": "^16.0.0",
+    "@angular-eslint/builder": "^16.0.0",
+    "@angular-eslint/eslint-plugin": "^16.0.0",
+    "@angular-eslint/eslint-plugin-template": "^16.0.0",
+    "@angular-eslint/schematics": "^16.0.0",
+    "@angular-eslint/template-parser": "^16.0.0",
+    "@angular/cli": "^16.0.0",
+    "@angular/compiler-cli": "^16.0.0",
+    "@angular/language-service": "^16.0.0",
+    "@badeball/cypress-cucumber-preprocessor": "^18.0.0",
     "@cypress/schematic": "^2.5.0",
     "@cypress/webpack-preprocessor": "^5.17.0",
-    "@sinonjs/fake-timers": "^9.1.2",
+    "@sinonjs/fake-timers": "^11.0.0",
     "@types/chai": "^4.1.7",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^12.11.1",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.2.5",
     "@types/sinon": "^10.0.13",
-    "@typescript-eslint/eslint-plugin": "^5.48.1",
-    "@typescript-eslint/parser": "^5.48.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
     "cassandra-driver": "^4.6.3",
     "chai": "^4.2.0",
     "cypress": "^12.8.1",
-    "eslint": "^8.12.0",
+    "eslint": "^8.39.0",
     "karma": "^6.4.1",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage": "~2.0.3",
+    "karma-coverage": "^2.2.0",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
-    "mocha": "^9.2.0",
+    "mocha": "^10.2.0",
     "process": "^0.11.10",
-    "sinon": "^11.1.1",
-    "sonarqube-scanner": "^2.8.2",
+    "sinon": "^15.1.0",
+    "sonarqube-scanner": "^3.0.1",
     "ts-loader": "^9.4.2",
-    "typescript": "^4.8.3",
+    "typescript": "^5.0.4",
     "util": "^0.12.4"
   },
   "cypress-cucumber-preprocessor": {
